:recycle: (typescript) migrated ManagePayees and LoadBackupModal files
