+++ packages/desktop-client/src/components/payees/PayeeTable.tsx
@@ -12,7 +12,7 @@ import { type PayeeEntity } from 'loot-core/src/types/models';
 
 import { useSelectedItems } from '../../hooks/useSelected';
 import { View } from '../common/View';
-import { Table, type TableNavigator } from '../table';
+import { Table } from '../table';
 
 import { PayeeTableRow } from './PayeeTableRow';
 
@@ -23,7 +23,6 @@ type PayeeWithId = PayeeEntity & Required<Pick<PayeeEntity, 'id'>>;
 type PayeeTableProps = {
   payees: PayeeWithId[];
   ruleCounts: Map<PayeeWithId['id'], number>;
-  navigator: TableNavigator<PayeeWithId>;
 } & Pick<
   ComponentProps<typeof PayeeTableRow>,
   'onUpdate' | 'onViewRules' | 'onCreateRule'
@@ -32,53 +31,46 @@ type PayeeTableProps = {
 export const PayeeTable = forwardRef<
   ComponentRef<typeof Table<PayeeWithId>>,
   PayeeTableProps
->(
-  (
-    { payees, ruleCounts, navigator, onUpdate, onViewRules, onCreateRule },
-    ref,
-  ) => {
-    const [hovered, setHovered] = useState(null);
-    const selectedItems = useSelectedItems();
+>(({ payees, ruleCounts, onUpdate, onViewRules, onCreateRule }, ref) => {
+  const [hovered, setHovered] = useState(null);
+  const selectedItems = useSelectedItems();
 
-    useLayoutEffect(() => {
-      const firstSelected = [...selectedItems][0] as string;
-      if (typeof ref !== 'function') {
-        ref.current.scrollTo(firstSelected, 'center');
-      }
-      navigator.onEdit(firstSelected, 'select');
-    }, []);
+  useLayoutEffect(() => {
+    const firstSelected = [...selectedItems][0] as string;
+    if (typeof ref !== 'function') {
+      ref.current.scrollTo(firstSelected, 'center');
+    }
+  }, []);
 
-    const onHover = useCallback(id => {
-      setHovered(id);
-    }, []);
+  const onHover = useCallback(id => {
+    setHovered(id);
+  }, []);
 
-    return (
-      <View style={{ flex: 1 }} onMouseLeave={() => setHovered(null)}>
-        <Table
-          ref={ref}
-          items={payees}
-          navigator={navigator}
-          renderItem={({ item, editing, focusedField, onEdit }) => {
-            return (
-              <PayeeTableRow
-                payee={item}
-                ruleCount={ruleCounts.get(item.id) || 0}
-                selected={selectedItems.has(item.id)}
-                editing={editing}
-                focusedField={focusedField}
-                hovered={hovered === item.id}
-                onHover={onHover}
-                onEdit={onEdit}
-                onUpdate={onUpdate}
-                onViewRules={onViewRules}
-                onCreateRule={onCreateRule}
-              />
-            );
-          }}
-        />
-      </View>
-    );
-  },
-);
+  return (
+    <View style={{ flex: 1 }} onMouseLeave={() => setHovered(null)}>
+      <Table
+        ref={ref}
+        items={payees}
+        renderItem={({ item, editing, focusedField, onEdit }) => {
+          return (
+            <PayeeTableRow
+              payee={item}
+              ruleCount={ruleCounts.get(item.id) || 0}
+              selected={selectedItems.has(item.id)}
+              editing={editing}
+              focusedField={focusedField}
+              hovered={hovered === item.id}
+              onHover={onHover}
+              onEdit={onEdit}
+              onUpdate={onUpdate}
+              onViewRules={onViewRules}
+              onCreateRule={onCreateRule}
+            />
+          );
+        }}
+      />
+    </View>
+  );
+});
 
 PayeeTable.displayName = 'PayeeTable';
