+++ packages/loot-core/src/types/models/payee.d.ts
@@ -4,6 +4,6 @@ export interface PayeeEntity {
   id: string;
   name: string;
   transfer_acct?: AccountEntity['id'];
-  favorite?: boolean;
+  favorite?: 1 | 0;
   tombstone?: boolean;
 }
