+++ upcoming-release-notes/3507.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+TypeScript: migrated `ManagePayees` and `LoadBackupModal`.
